+++ packages/basic/src/apis/index.ts
@@ -3,3 +3,5 @@ export { initService as initConfigurationService } from "./configuration";
 export { initService as initIoService } from './io'
 export { initService as initLowauthService } from "./lowauth";
 export { initService as initProcessService } from "./process";
+export { initService as initProcessV2Service } from "./processV2";
+export { initService as initSystemProcessV2Service } from "./system/processV2";
