+++ packages/basic/src/apis/processV2/index.ts
@@ -1,0 +1,12 @@
+import { createService } from '../../utils/create';
+import api from './api.json';
+
+type Service<API> = {
+  [key in keyof API]: (params?: any) => Promise<any>
+};
+
+const initService = () => createService(api) as Service<typeof api>;
+
+export {
+    initService,
+};
