+++ packages/basic/src/init/process/index.ts
@@ -1,15 +1,19 @@
-import { initService as processV2Service } from "./processV2";
-import { initService as systemProcessV2Service } from "./system/processV2";
+import { initProcessV2Service, initSystemProcessV2Service } from '../../apis';
 import Global from "../../global";
+import processService from "./processService";
 
 function initProcess() {
   /**
    * 流程接口注册
    */
-  Global.prototype.$processV2 = processV2Service();
-  Global.prototype.$systemProcessV2 = systemProcessV2Service();
+  Global.prototype.$process = processService;
+  Global.prototype.$processV2 = initProcessV2Service();
+  Global.prototype.$systemProcessV2 = initSystemProcessV2Service();
 }
 
 export {
-  initProcess
+  initProcess,
+  processService
 }
+
+export { processPorts } from './processPorts';
