--- packages/basic/src/init/process/processV2/index.ts
@@ -1,8 +1,0 @@
-import { createService } from '../../../utils/create';
-import api from './api.json';
-
-const initService = () => createService(api);
-
-export {
-    initService,
-};
